refactor(Order): Order에서 Table 의존 로직은 통합계층으로 옮김
